requirements.txt
ollama==0.6.0
rich==14.2.0
numpy==2.3.4
python-dotenv==1.2.1
